docs(robovm): add screenshots, README update, and run-sim.sh
- Capture 5 screenshots: login, home, product, cart, receipt
- Rewrite README.md with screenshots, UIKitDom docs, updated structure
- Add run-sim.sh build/deploy script with --no-clean and --deps flags
- Minor view cleanups from UIKitDom refactoring

+++ fontes/br.com.wdc.shopping/br.com.wdc.shopping.view.robovm/src/main/java/br/com/wdc/shopping/view/robovm/impl/PurchasesPanelViewRoboVM.java
@@ -62,7 +62,7 @@ public class PurchasesPanelViewRoboVM extends AbstractViewRoboVM<PurchasesPanelP
         this.contentSlot.accept(this.state.purchases, this.itemViewList);
 
         // Update scroll content size
-        int count = this.itemViewList.size();
+        double count = this.itemViewList.size();
         scrollView.setContentSize(new CGSize(343, count * 44));
 
         // Update pagination label
